+++ setupSlave.sh
@@ -1,6 +1,6 @@
 # nafsdm
 # help script for setting up nafsdm on all slaves
-# Copyright Vilhelm Prytz 2017
+# Copyright Vilhelm Prytz 2018
 # https://github.com/mrkakisen/nafsdm
 
 # check if user is root or not
@@ -9,7 +9,8 @@ if [[ $EUID -ne 0 ]]; then
   exit 1
 fi
 
-#  DL_VERSION will be changed at the time of update
+CLONE_URL="https://github.com/MrKaKisen/nafsdm.git"
+CLONE_BRANCH="development"
 DL_URL="https://github.com/MrKaKisen/nafsdm/archive/"
 GITHUB_DIR="slave-nafsdm"
 HOME_DIR="/home/slave-nafsdm"
@@ -49,39 +50,67 @@ fi
 echo "* Fetching information about latest version.."
 LATEST_VERSION=$(curl https://raw.githubusercontent.com/MrKaKisen/nafsdm/master/version.txt)
 
-# select version
-echo "* Please select your version. Type in the version number or type 'latest' for latest version."
-echo -n "* Version: "
-read VERSION_USER
+# commit updates
+echo "* Developers only: Would you like to enable incremental commit updates and use development branch only?"
+echo "* Warning: This is a developer function, do not use in a production environment."
+echo -n "* Enable? (y/n): "
+read DEV_IC_CONFIRM
 
-if [ "$VERSION_USER" == "latest" ]; then
-  echo -n "* Confirm? (y/n): "
-  read CONFIRM
-  if [ "$CONFIRM" == "y" ]; then
-    DL_VERSION="$LATEST_VERSION"
+if [ "$DEV_IC_CONFIRM" == "y" ]; then
+  if [ "$OPERATINGSYS" == "centos" ]; then
+    yum install git -y
+  elif [[ "$OPERATINGSYS" == "debian" ]] || [[ "$OPERATINGSYS" == "ubuntu" ]] ; then
+    apt-get install git -y
   else
-    echo "* Aborting.."
+    echo "* Invalid operating system."
     exit 1
+  fi
+
+  cd /tmp
+  git clone -b $CLONE_BRANCH $CLONE_URL
+elif [ "$DEV_IC_CONFIRM" == "n" ]; then
+  echo "* Skipping.."
+  # select version
+  echo "* Please select your version. Type in the version number or type 'latest' for latest version."
+  echo -n "* Version: "
+  read VERSION_USER
+
+  if [ "$VERSION_USER" == "latest" ]; then
+    echo -n "* Confirm? (y/n): "
+    read CONFIRM
+    if [ "$CONFIRM" == "y" ]; then
+      DL_VERSION="$LATEST_VERSION"
+    else
+      echo "* Aborting.."
+      exit 1
+    fi
+  else
+    echo -n "* Confirm? If version doesn't exist, script will fail. (y/n): "
+    read CONFIRM
+    if [ "$CONFIRM" == "y" ]; then
+      DL_VERSION="$VERSION_USER"
+    else
+      echo "* Aborting.."
+      exit 1
+    fi
   fi
 else
-  echo -n "* Confirm? If version doesn't exist, script will fail. (y/n): "
-  read CONFIRM
-  if [ "$CONFIRM" == "y" ]; then
-    DL_VERSION="$VERSION_USER"
-  else
-    echo "* Aborting.."
-    exit 1
-  fi
+  echo "* Aborting.."
+  exit 1
 fi
 
 echo "* Required packages installed!"
 echo "* Downloading nafsdm & installing.."
 
 # download in temp dir
+if [ "$DEV_IC_CONFIRM" == "n" ]; then
+  cd /tmp
+  wget $DL_URL$DL_VERSION.tar.gz -O nafsdm.tar.gz
+  tar -zxvf nafsdm.tar.gz
+  mv nafsdm-* nafsdm
+fi
+
 cd /tmp
-wget $DL_URL$DL_VERSION.tar.gz -O nafsdm.tar.gz
-tar -zxvf nafsdm.tar.gz
-mv nafsdm-* nafsdm
 
 useradd $USER
 # debian and ubuntu doesn't create its home dir automatically, unlike centos
@@ -102,6 +131,17 @@ cp /tmp/nafsdm/systemconfigs/nafscli /usr/bin/nafscli
 
 chmod +x /home/slave-nafsdm/start.py
 chmod +x /usr/bin/nafscli
+
+# dev set version
+if [ "$DEV_IC_CONFIRM" == "y" ]; then
+  cd /tmp/nafsdm
+  COMMIT_HASH=$(git log -n 1 development | sed -n '1p' | cut -c8-14)
+  echo "version = \"$COMMIT_HASH-dev\"" > /home/slave-nafsdm/pythondaemon/version.py
+
+  # enable IC mode and change branch to master
+  sed -i '10s/.*/github_branch = development/' /home/slave-nafsdm/config.conf
+  sed -i '12s/.*/incrementalCommitVersions = True/' /home/slave-nafsdm/config.conf
+fi
 
 echo "* Installed. Cleanup.."
